fix tests

+++ db/db_test2.cc
@@ -3225,7 +3225,7 @@ TEST_F(DBTest2, MultiDBParallelOpenTest) {
 
   // Verify empty DBs can be created in parallel
   std::vector<std::thread> open_threads;
-  std::vector<DB*> dbs{kNumDbs, nullptr};
+  std::vector<DB*> dbs{static_cast<unsigned int>(kNumDbs), nullptr};
   options.create_if_missing = true;
   for (int i = 0; i < kNumDbs; ++i) {
     open_threads.emplace_back(
